Remove setting for Wallpaper
This has never made sense to me and really question why is even in AOSP. I've never
seen any user or know anyone that goes into Settings to change their wallpaper. Most
if not all launchers have an option to change the wallpaper and that's the method
that users like and use.

Change-Id: I7b31c428218c50db6a5ffa8227d1ff47e6000713